+++ src/libmbedcrypto.c
@@ -88,7 +88,7 @@ SHACTX sha1_init(void)
     return ctx;
 }
 
-void sha1_update(SHACTX c, const void *data, unsigned long len)
+void sha1_update(SHACTX c, const void *data, size_t len)
 {
     mbedtls_md_update(c, data, len);
 }
@@ -100,7 +100,7 @@ void sha1_final(unsigned char *md, SHACTX c)
     SAFE_FREE(c);
 }
 
-void sha1(const unsigned char *digest, int len, unsigned char *hash)
+void sha1(const unsigned char *digest, size_t len, unsigned char *hash)
 {
     const mbedtls_md_info_t *md_info =
         mbedtls_md_info_from_type(MBEDTLS_MD_SHA1);
@@ -122,7 +122,7 @@ static mbedtls_md_type_t nid_to_md_algo(int nid)
     return MBEDTLS_MD_NONE;
 }
 
-void evp(int nid, unsigned char *digest, int len,
+void evp(int nid, unsigned char *digest, size_t len,
         unsigned char *hash, unsigned int *hlen)
 {
     mbedtls_md_type_t algo = nid_to_md_algo(nid);
@@ -170,7 +170,7 @@ EVPCTX evp_init(int nid)
     return ctx;
 }
 
-void evp_update(EVPCTX ctx, const void *data, unsigned long len)
+void evp_update(EVPCTX ctx, const void *data, size_t len)
 {
     mbedtls_md_update(ctx, data, len);
 }
@@ -216,7 +216,7 @@ SHA256CTX sha256_init(void)
     return ctx;
 }
 
-void sha256_update(SHA256CTX c, const void *data, unsigned long len)
+void sha256_update(SHA256CTX c, const void *data, size_t len)
 {
     mbedtls_md_update(c, data, len);
 }
@@ -228,7 +228,7 @@ void sha256_final(unsigned char *md, SHA256CTX c)
     SAFE_FREE(c);
 }
 
-void sha256(const unsigned char *digest, int len, unsigned char *hash)
+void sha256(const unsigned char *digest, size_t len, unsigned char *hash)
 {
     const mbedtls_md_info_t *md_info =
         mbedtls_md_info_from_type(MBEDTLS_MD_SHA256);
@@ -270,7 +270,7 @@ SHA384CTX sha384_init(void)
     return ctx;
 }
 
-void sha384_update(SHA384CTX c, const void *data, unsigned long len)
+void sha384_update(SHA384CTX c, const void *data, size_t len)
 {
     mbedtls_md_update(c, data, len);
 }
@@ -282,7 +282,7 @@ void sha384_final(unsigned char *md, SHA384CTX c)
     SAFE_FREE(c);
 }
 
-void sha384(const unsigned char *digest, int len, unsigned char *hash)
+void sha384(const unsigned char *digest, size_t len, unsigned char *hash)
 {
     const mbedtls_md_info_t *md_info =
         mbedtls_md_info_from_type(MBEDTLS_MD_SHA384);
@@ -323,7 +323,7 @@ SHA512CTX sha512_init(void)
     return ctx;
 }
 
-void sha512_update(SHA512CTX c, const void *data, unsigned long len)
+void sha512_update(SHA512CTX c, const void *data, size_t len)
 {
     mbedtls_md_update(c, data, len);
 }
@@ -335,7 +335,7 @@ void sha512_final(unsigned char *md, SHA512CTX c)
     SAFE_FREE(c);
 }
 
-void sha512(const unsigned char *digest, int len, unsigned char *hash)
+void sha512(const unsigned char *digest, size_t len, unsigned char *hash)
 {
     const mbedtls_md_info_t *md_info =
         mbedtls_md_info_from_type(MBEDTLS_MD_SHA512);
@@ -377,7 +377,7 @@ MD5CTX md5_init(void)
 }
 
 
-void md5_update(MD5CTX c, const void *data, unsigned long len) {
+void md5_update(MD5CTX c, const void *data, size_t len) {
     mbedtls_md_update(c, data, len);
 }
 
@@ -397,7 +397,7 @@ int ssh_kdf(struct ssh_crypto_struct *crypto,
                              key_type, output, requested_len);
 }
 
-HMACCTX hmac_init(const void *key, int len, enum ssh_hmac_e type)
+HMACCTX hmac_init(const void *key, size_t len, enum ssh_hmac_e type)
 {
     HMACCTX ctx = NULL;
     const mbedtls_md_info_t *md_info = NULL;
@@ -446,7 +446,7 @@ error:
     return NULL;
 }
 
-void hmac_update(HMACCTX c, const void *data, unsigned long len)
+void hmac_update(HMACCTX c, const void *data, size_t len)
 {
     mbedtls_md_hmac_update(c, data, len);
 }
